Revert "Move tests for moved methods"
This reverts commit 8267bf09db86d7c4a721129fb57066ded2a2b4fb.
